codecheck

+++ services/permission_standard/distributedpermissionmanagerservice/test/unittest/soft_bus_channel_test/remote_command_executor_test.cpp
@@ -180,7 +180,7 @@ HWTEST_F(RemoteCommandExecutorTest, RemoteCommandExecutor_Construct_002, TestSiz
     time_t end = TimeUtil::GetTimestamp();
     PERMISSION_LOG_DEBUG(LABEL,
         "RemoteCommandExecutor_Construct_002: ellapsed %{public}ld second(s) from %{public}ld to %{public}ld",
-        (long) end - begin, (long) begin, (long) end);
+        (long) end - (long)begin, (long) begin, (long) end);
 
     PERMISSION_LOG_DEBUG(LABEL, "start: %{public}ld", (long) begin);
     PERMISSION_LOG_DEBUG(LABEL, "end: %{public}ld", (long) end);
